Bug fix: Crash when shooting tank with Launcher
On Eurocorp mail 15 mission, we deal with a tank at the beginning.
If the tank goes up the ramp, and we shoot iw with Launcher,
the game often crashes.

This hopefully fixes the issue. Though I can't tell whether it's a
proper fix, I can't fully understand the code. If a pointer which
was invalid here points to a thing, then this is a proper workaround.

To properly fix this, we would really have to know why the value written
there is invalid even though it should be a pointer. But this is out
of reach in current state of the project.