swars: Closing colour tables into pixmap, generation
Pixmap is now a struct that contain all colour tables. All
references in C were updated.

Also, '-T' option for generating colour tables was implemented, and
the option for generating from fade table was improved. It is still
not recommended to use these options - original colour tables are
hand-crafted, and the manual changes are lost on such generation.