swars: Removed one of the ways of FPS limiting
The FPS was guarded by two completely different wait mechanisms,
which sometimes caused confusing effects.

Now the frame guarding is still doubled in some cases,
but at least there is the same limiting code used.

The duplicated limiting needs to be removed, but that is
not in scope of this patch.