swars: Fixed out of bounds write in scanner
Recent rewrite of scanner related functions made it possible
to scale the scanner with resolution. Due to that scaling not
being restricted, it was possible to set height larger than
possible amout of lines due to static array in use.

This change restrict scanner minimap height.
Also a function was remade, not very important one.