swars: Fix playing samples on tank turret rotation
Besides being played at coords 0,0 instead of where the tank is,
the issue with the original code was that it played the stopping
sample constantly. Now it only plays the stop sample once after
each rotation sample. Also, the rotation sample is now played
under modified conditions - requires larger rotation speed,
and requires the speed to increase.

The rotation sample is also no longer looping, because the way it
sounds isn't good for looping - it sounds way better on single play.