swars: Remade read_user_settings()
The existence of key.dat file was causing no sound in menu. This
is because the game tried to read settings from there, but they
were not available.

Now the setting reading was remade, but also `keys.dat` stopped
being a default settings file. The file is expected to only
store a copy of last mortal game encryption salt.