swars: Work on refreshing old Thing format part 6
This finishes review of all the currently known info about fields within
the old Thing struct. All the known and suspected fields are now properly
copied.

Which doesn't mean we have it all. Especially not having the cyborg
mods has profound impact on the gameplay. But other fields are also
missing, and some are uncertain.