swars: Switched implementation of fix_thing_command_indexes()
This is a big change - instead of a big switch for selecting commands
based on types, our new command definitions are used.

If there are any mistakes in our command definitions, this change
may introduce bugs - need to be thoroughly tested!

But if there are no mistakes, then this should fix few ocmmands which
were misbehaving.