Minor Text fixes
These are fixes for errors inherited from the original game. Deadly Harvest is actually mission 43, not 40, as confirmed by a) every ALLTEXT.WAD other than the English one, and b) the official strategy guide. Secondly Bahrain is a country, the name of the actual city (in Bahrain) is Al Manamah. Weirdly this is corrected in ALL the demos of the game, but wrong in the actual game.