swars: Finally unified use of init_game()
The way it is done has issues - the level is loaded
while black screen is shown rather than loading screen.
So either the init_game() should be separated from
any colour computations, or something else should
be drawn while the function works.

Not a big deal - the black screen only lasts a split
second, especially on recent hardware. But still..