swars: Implemented face_is_blocking_walk() using flags, moved update
Instead of re-computing steepness, face_is_blocking_walk() is now just
checkinf face flags. And to re-compute that flag, the steepness
computation has been moved to a better file, and a function which
updates the flag for all faces was added.

Currently, the update function is unused.

Also removed things debug mode, mistakenly enabled in one
of previous commits.