swars: Workaround vehicle blockades
It was possible for two vehicles to block on each other. with one
blocking due to proximity of another, and the other blocking on
flags of the TNode. This was happening when both vehicles were on
the crossroads.

This change is probably not a best fix to this issue, but it
workarounds it enough for the blockades to cease existing.

Maybe instead, while checking TNodes, links to further TNodes
on all sides should be taken into consideration?