swars: Several unifications around mission state slots
The game has a number of slots to keep state of missions.
They were updated in varoious ways depending on place.
This change unifies that a bit, though more changes are
needed for that.

This also introduces a clear function which removes old data
while starting a new game. This fizxes an original game
bug where the slots could be filled with levels from previous
games, making it impossible to track completion of new
missions.