swars: Enable domes being passable when opened
This was prevented due to two bugs in the game code.
Original levels were adjusted for this bug, so an update
to Zealot campaign was required to make it work properly.

Now as levels are updated to mind the domes, this patch
enables the code fix which makes entering domes possible.