build: After building bflibrary, install it to output folder
The static libraries should be installed before use. This ensures
that any headers generated while building will be with the rest,
and private headers will get dropped.

We will want to expand the use of the config headers generated by
autotools, and this step is required for that.

This change ensures that bflibrary is installed, and then uses
the installed version while building the swars application.