swars: Standard WScreen manipulation fucts used in shadows gen
The shadows generation function used some direct commands to manipulate
working screen buffer. But we already have proper functions which
do that. This patch removes the custom code and uses the generic
functions instead.

It still leaves the same screen buffer in use - which is not a problem
now, but at some point we will want to stop using a standard CPU
buffer as screen buffer.