swars: Improvements in panel drawing
Agents panels are darkened if the agent is executing
predefined commands list rather than being controlled
by player.
In most levels, panel was disabled when executing commands;
but not in all. This provides clear indication the player
is not in control. Example is `-m 1,27`.

Health/mood overlay is now drawn only if the panel below
is drawn first. Disabling the panel sprite, disables its
content automatically.

Weapons are still drawn separately from the panel, but
the function is now informed whether the panel was drawn
or not. So they also can no longer appear without the
panel being visible.