swars: Introduced get_thing_position_mapcoords()
The thing coordinates are a mess. They're in precise coords, unless its
Y coord which is shifted 3 bits to that, unless it's a building because
then it isn't shifted. And tank turrets have relative coords, for which
you have to find the base.

We cannot fix the mess. But to manage it, we can introduce a function
which will just return normal coordinates, always with the same definition.

Also named two lookup colors.