swars: refactored flashy_draw_purple_text_box() to be more dynamic
It can now cap the scrollbar hole while keeping it in stored
background. This allows enabling and disabling the scrollbar
in lifetime of a box.

Holes are capped only after drawing finishes, which leads to
not neccessarily wanted effects during animation. This may
require another change in the future - currenlty I do not have
an idea on how to fix that without ugly patching.