swars: Enable the backward playing cyborg mod out animations
In order to show a mod being removed from cyborg view, it was
neccessary to play the animation in reverse. Otherwise, the
progressing animation would reveal backgroun behind the mod,
making it impossible to create an animation which fits any
background.

This patch requires corresponding change to swars-gfx package,
where the backward-playing animations should be provided.