swars: Include lang dir in game_dirs[]
This breaks the original concept behind it, where it would
contain places which map to language path only if CD is in
use. But we will no longer use the CD, so this feature will
get removed over time.

Instead, we want the game on disk to contain all languages
and allow to easily switch between them.