config: Remove incorrect video from Briefing of 'The Dead Zone'
The miss001.ini carried a mistake since the original game. A video was assigned to briefing of mission 'The Dead Zone', which was showing a different level. This change Removes that video. It is suspected that this level was not supposed to have a video briefing, as the briefing text suggests lack of knowledge about what is happening within the city.