bflibrary: Hopefully now really proper way of keeping scaled sizes
The patch "Proper way for keeping rescaled sprite sizes" tured out
to be overcomplicated, and problematic due to not keeping bounds
correctly.

A symptom of the issue was that at level boundaries, parts of sprites
were visible outside of map area.

Instead of trying to fix the sizes array after generation, this
change makes sure the generation is done correctly, to keep the sizes.

The core of the change is to add 0x7fff to high-precision values,
resembling addition of 0.5 to minimize float-to-integer rounding
errors.